# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aluf Messenger API
  description: |
    REST API для мессенджера нового поколения **Aluf**.
    
    ## Аутентификация
    Все защищённые эндпоинты требуют JWT Bearer токен в заголовке `Authorization`.
    Токены генерируются через `/v1/auth/login` → `/v1/auth/verify`.
    
    ## Лимиты
    - Rate limit: 100 req/min (глобальный), 20 req/min (auth)
    - Максимальный размер сообщения: 4096 символов
    - Максимальный размер файла: 4 ГБ (8 ГБ для Premium)
    - Максимум участников группы: 200 000
    
    ## WebSocket
    Real-time подключение через защищённый WebSocket endpoint среды, в которой развёрнут сервис (конкретный URL не фиксируется в этой спецификации).
  version: 1.0.0
  contact:
    name: Aluf Team
  license:
    name: MIT
servers:
  - url: https://example.com/v1
    description: Example base URL (replace with your deployment)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация по номеру телефона
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: "+79001234567"
                displayName:
                  type: string
                  example: Алиса
      responses:
        '200':
          description: OTP отправлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        '429':
          description: Слишком много запросов
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход — запрос OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: "+79001234567"
      responses:
        '200':
          description: OTP отправлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /auth/verify:
    post:
      tags: [Auth]
      summary: Подтверждение OTP — получение токенов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId, code]
              properties:
                requestId:
                  type: string
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: "123456"
                deviceInfo:
                  type: object
                  properties:
                    platform:
                      type: string
                    model:
                      type: string
                    os:
                      type: string
      responses:
        '200':
          description: Успешная верификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Неверный или просроченный OTP
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Невалидный refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход из текущей сессии
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Сессия завершена
  /auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Настройка двухфакторной аутентификации (TOTP)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: QR-код и секрет для настройки
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qrCodeUrl:
                    type: string
                  backupCodes:
                    type: array
                    items:
                      type: string
  /auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Подтверждение 2FA кода
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: 2FA подтверждено

  # === BOTS ===
...